Projects/fin_etl/requirements.txt
google-cloud-firestore==2.16.*
prophet==1.1.*
pandas==2.2.*